+++ muscopy/qpi.py
@@ -349,6 +349,8 @@ def _get_phase_noise(
 ) -> NDArray:
     """Calculates the phase noise in a given array.
 
+    Here, dc_intensity is the photon number of DC components, as the shot noise is proportional to the square root of the photon number.
+
     Args:
         visibility (NDArray): The visibility of the object.
         aperturesize (int): The size of the aperture.
@@ -389,18 +391,25 @@ def calc_visibility(
 def calc_phase_noise(
     hologram: NDArray,
     params: QPIParameters,
+    fullwell: float,
+    quantum_eff: float,
+    bit_depth: int,
 ) -> NDArray:
     """Calculates the phase noise in a given array.
 
     Args:
         hologram (NDArray): The hologram of the object.
         params (mus.QPIParameters): The parameters of the QPI.
+        fullwell (float): The full well capacity of the sensor.
+        quantum_eff (float): The quantum efficiency of the sensor
+        bit_depth (int): The bit depth of the sensor.
 
     Returns:
         NDArray: The phase noise
     """
     dc, ac = _get_dc_ac(hologram, params)
     visibility = _get_visibility(dc, ac)
-    phase_noise = _get_phase_noise(visibility, params.aperturesize, xp.abs(dc), params.img_shape)
+    dc_factor = fullwell / quantum_eff / (2**bit_depth)
+    phase_noise = _get_phase_noise(visibility, params.aperturesize, dc_factor * xp.abs(dc), params.img_shape)
 
     return phase_noise